ai_backend/requirements.txt
flask>=3.0.0
streamlit>=1.37.0
ultralytics>=8.3.0
opencv-python>=4.10.0
numpy>=1.26.0
torch>=2.4.0
deep-sort-realtime>=1.3.2
requests>=2.32.0
python-dotenv>=1.0.1
face-recognition>=1.3.0
